+++ tests/functional/test__cmdline.py
@@ -25,7 +25,7 @@ class TestCmdlineEditing(unittest.FunctionalTestCase):
 
     def setUp(self):
         super().setUp()
-        self.view.assign_syntax('Packages/NeoVintageous/res/Command-line mode.sublime-syntax')
+        self.view.settings().set('_nv_ex_mode', True)
         reset_cmdline_completion_state()
 
     # TODO [refactor] Into usable run test command-line mode command via feed.